added rlock and minor changes to code
